merge: data.compare and data.source to data.py
