+++ app.py
@@ -13,8 +13,7 @@ from linebot.models import (
     MessageEvent, TextMessage, TextSendMessage,
 )
 
-import data.source
-import data.compare
+import data
 
 app = Flask(__name__)
 
@@ -38,7 +37,7 @@ def callback():
 
 @handler.add(MessageEvent, message=TextMessage)
 def handle_message(event):
-    similarty_result = data.compare.questions(word=event.message.text, samples=data.source.from_csv_file('./data.csv'))
+    similarty_result = data.compare_questions(word=event.message.text, samples=data.from_csv_file('./data.csv'))
     line_bot_api.reply_message(
         event.reply_token,
         TextSendMessage(text=similarty_result.to_json()))
